tools: mailsend: Begin revamp of loader app
The bootstrapping is going to be split in to three stages.
Stage 0 is going to be the absolute smallest binary it can be.
Stage 1 and stage 2 will be a back to back loader, with stage 2
being the full featured loader in the end and a testbed for new
features.

Initial commit of stage 1 and 2 loaders. Stage 1 requires
a custom crt0 that is specific for it.

Stage 2 currently supports self-write to App 0 slot, load/run from
parpart for programs from either 0x8000 or 0x4000 (limited to
16kbyte page), reboot (this is still kind of broken as it has
always been), and poweroff.