mslib: crt0-slot4: Initial commit
Still rough, but is a decent outline of whats needed for a crt0
for apps being launched from dataflash.

Note that, I think I'm going to abandon the "try to use as much
of the stock firmware as possible" idea. I just don't think that
is a valid goal considering how much of the firmware has random
bits of memory used. Probably going to push forward with crt0's
implementing a custom ISR as Fyber did. Might be worth
investigating the actual MS hardware further to see what the bus
pins do in this case during an IRQ