tools: Mass rename/separation from mailsend
The Mailstation side was renamed to msloader. Each of the stages
were renamed a more human friendly set of names.
LDR, msloader-LDR, and msloader.

The PC side was moved to src/host/mailtransfer and was refreshed
to support specifying LPT port address to accommodate PCIe LPT
interfaces that put the ports at non-traditional IO addrs. This
was also updated to add in maildump functionality, combining the
two tools in to a single tool (similar to what msloader has now).

The host/mailtransfer folder includes binary dll and .exe. It is
not built as a part of a global 'make' operation. Requires MinGW.
Still only supports Windows at this time.