msloader: Refactor, data xfer between MSs, bootstrap
Break out the single msloader.c file to multiple pieces to segment
and more easily manage. Each file now is responsible for a single
objective.

Add ability to send dataflash slots to another MailStation. Also
add ability for msloader to bootstrap another MailStation. This
takes the compile msloader-LDR trampoline and LDR code and
builds it in to msloader. Instructions for bootstrapping another
unit are provided on-screen. These are how to enter the hex viewer,
how to enter the hex editor, what bytes to add to enable app 0,
and what bytes to enter to install LDR to app 0. After this,
msloader will send the trampoline, followed by itself, to the
virgin MailStation which can run from RAM or install itself to
app 0.

Bump to 0.24